ICES_SMAC/src/requirements.txt
absl_py==1.4.0
cloudpickle==1.2.2
gym==0.15.3
matplotlib==3.6.3
numpy==1.22.4
pygame==2.3.0
pyro_ppl==1.8.4
PySC2==3.0.0
PyYAML==5.4.1
s2clientprotocol==5.0.11.89720.0
setproctitle==1.3.2
SMAC==1.0.0
torch==1.13.1+cu117
wandb==0.15.8
